+++ src/database/models/setting.js
@@ -1,0 +1,25 @@
+const { DataTypes } = require('sequelize');
+
+module.exports = (sequelize) => {
+  const Setting = sequelize.define('Setting', {
+    id: {
+      type: DataTypes.INTEGER,
+      autoIncrement: true,
+      primaryKey: true,
+    },
+    key: {
+      type: DataTypes.STRING,
+      allowNull: false,
+      unique: true,
+    },
+    value: {
+      type: DataTypes.STRING(1024),
+      allowNull: false,
+    },
+  }, {
+    tableName: 'settings',
+    timestamps: true,
+  });
+
+  return Setting;
+};